refactor: refactor and simplify db seed

+++ backend/src/main/java/net/tamasnovak/repositories/ApplicationRepository.java
@@ -3,6 +3,6 @@ package net.tamasnovak.repositories;
 import net.tamasnovak.entities.application.Application;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApplicationRepository extends JpaRepository<Application, Integer> {
+public interface ApplicationRepository extends JpaRepository<Application, Long> {
 
 }
